[cleanup] Move artefact related methods to a dedicated module
